similar_asserts

Macro assert_eq

macro_rules! assert_eq {
    ($left_label:ident: $left:expr, $right_label:ident: $right:expr $(,)?) => { ... };
    ($left_label:ident: $left:expr, $right_label:ident: $right:expr, $($arg:tt)*) => { ... };
    ($left:expr, $right:expr $(,)?) => { ... };
    ($left:expr, $right:expr, $($arg:tt)*) => { ... };
}

Asserts that two expressions are equal to each other (using PartialEq).

On panic, this macro will print the values of the expressions with their Debug or ToString representations with a colorized diff of the changes in the debug output. It picks Debug for all types that are not strings themselves and ToString for str and String.

Like assert!, this macro has a second form, where a custom panic message can be provided.

use similar_asserts::assert_eq;
assert_eq!((1..3).collect::<Vec<_>>(), vec![1, 2]);